backend/app/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
openpyxl
pdfplumber
pytesseract
Pillow
python-multipart
python-dotenv
pytest
httpx
pytest-cov
prometheus-client
pydantic-settings
redis
rq
croniter
scikit-learn
joblib
statsmodels
passlib[bcrypt]
python-jose[cryptography]
email-validator
boto3
sqlalchemy>=2.0
psycopg2-binary
sentry-sdk
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
psutil
